+++ pkg/local_object_storage/shard/list.go
@@ -14,6 +14,11 @@ import (
 // Cursor is a type for continuous object listing.
 type Cursor = meta.Cursor
 
+// NewCursor creates a Cursor positioned at the given container and object.
+func NewCursor(cnr cid.ID, obj oid.ID) *Cursor {
+	return meta.NewCursor(cnr, obj)
+}
+
 // ErrEndOfListing is returned from object listing with cursor
 // when storage can't return any more objects after provided
 // cursor. Use nil cursor object to start listing again.