+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Changelog for NeoFS Node
 - Optimized locking for reputation data collector (#3851)
 - Optimized local HEAD request execution (#3783)
 - Unpaid container's data is deleted now (#3691)
+- Policer iterates engine-level object list now instead of shard-level (#3862)
 
 ### Removed
 - `node.persistent_sessions.path` config option from SN config (#3846)